🧳 Migration to ESM

+++ plugins/tickets/src/class/TicketBuilder.ts
@@ -1,23 +1,26 @@
-import { console } from "@/controller/console";
-import { Database } from "@/controller/database";
-import { Discord } from "@/discord/base";
-import { ButtonBuilder } from "@/discord/base/CustomIntetaction";
-import { Error } from "@/discord/base/CustomResponse";
-import Ticket, { Event, History, TicketCategories, Message as TicketMessage, TicketType, User as UserTicket, Voice } from "@/entity/Ticket.entry";
-import { ActionDrawer } from "@/functions/actionDrawer";
+import { Database } from "@/controller/database.js";
+import { Discord } from "@/discord/base/index.js";
+import { ButtonBuilder } from "@/discord/base/CustomIntetaction.js";
+import { Error } from "@/discord/base/CustomResponse.js";
+import Template from "@/entity/Template.entry.js";
+import Ticket, { Event, History, TicketCategories, Message as TicketMessage, TicketType, User as UserTicket, Voice } from "@/entity/Ticket.entry.js";
+import { ActionDrawer } from "@/functions/actionDrawer.js";
 import { ActionRowBuilder, ButtonInteraction, ButtonStyle, ChannelType, CommandInteraction, EmbedBuilder, Message, ModalSubmitInteraction, OverwriteResolvable, PermissionsBitField, StringSelectMenuInteraction, TextChannel, User, codeBlock } from "discord.js";
-import { ClaimBuilder } from "./ClaimBuilder";
+import { ClaimBuilder } from "./ClaimBuilder.js";
 
 type Interaction = CommandInteraction<'cached'> | ModalSubmitInteraction<'cached'> | ButtonInteraction<'cached'> | StringSelectMenuInteraction<'cached'> | Message<true>
 
 const ticket = new Database<Ticket>({ table: 'Ticket' })
+const template = new Database<Template>({ table: 'Template' })
 
 export class TicketBuilder {
   private options!: TicketType
-  public embed!: EmbedBuilder | undefined
-  public buttons!: ActionRowBuilder<ButtonBuilder>[] | undefined
+  public embed?: EmbedBuilder
+  public buttons?: ActionRowBuilder<ButtonBuilder>[]
   private user!: User
-  private templateId: number | undefined
+  private channelId?: string
+  private ticketId?: number
+  private templateId?: number
   private readonly interaction: Interaction
   constructor ({ interaction }: { interaction: Interaction }) {
     this.interaction = interaction
@@ -43,8 +46,15 @@ export class TicketBuilder {
     }
   }
 
-  setOwner (id: string) { this.options.ownerId = id; return this }
+  setData(data: Ticket) { 
+    this.options = Object.assign(this.options, data)
+    this.ticketId = data.id
+    return this
+  }
+  setTicket (channelId: string) { this.channelId = channelId; return this }
   setTemplateId (id: number) { this.templateId = id; return this}
+  
+  setOwner (id: string) { this.options.ownerId = id; return this }
   setTitle (content: string) { this.options.title = content; return this }
   setDescription(content: string) { this.options.description = content; return this }
   setClosed(isClosed: boolean) { this.options.closed = isClosed ?? false; return this }
@@ -52,7 +62,6 @@ export class TicketBuilder {
   setCategory (category: TicketCategories) { this.options.category = category; return this }
   setUser (user: User) { this.user = user; return this }
 
-  setData(data: Ticket) { this.options = Object.assign(this.options, data); return this }
 
   addTeam (user: UserTicket) { this.options.team.push(user); return this }
   addUsers (user: UserTicket) { this.options.users.push(user); return this }
@@ -60,7 +69,7 @@ export class TicketBuilder {
   addHistory (content: History) { this.options.history.push(content); return this }
   addMessage (message: TicketMessage) { this.options.messages.push(message); return this }
 
-  private permissions (): OverwriteResolvable[] {
+  public permissions (): OverwriteResolvable[] {
     const { guild } = this.interaction
     const { team, users, ownerId } = this.options
     const permissionOverwrites: OverwriteResolvable[] = []
@@ -72,36 +81,13 @@ export class TicketBuilder {
       PermissionsBitField.Flags.ReadMessageHistory
     ]
 
-    permissionOverwrites.push({
-      id: guild.id,
-      deny: [PermissionsBitField.Flags.ViewChannel]
-    })
+    if (ownerId !== '' && ownerId !== undefined) permissionOverwrites.push({ id: ownerId, allow: permissions })
+    
+    permissionOverwrites.push({ id: guild.id, deny: [PermissionsBitField.Flags.ViewChannel] })
+    permissionOverwrites.push({ id: this.user.id, allow: permissions })
 
-    for (const user of team) {
-      permissionOverwrites.push({
-        id: user.id,
-        allow: permissions
-      })
-    }
-
-    for (const user of users) {
-      permissionOverwrites.push({
-        id: user.id,
-        allow: permissions
-      })
-    }
-
-    if (ownerId !== '' && ownerId !== undefined) {
-      permissionOverwrites.push({
-        id: ownerId,
-        allow: permissions
-      })
-    }
-
-    permissionOverwrites.push({
-      id: this.user.id,
-      allow: permissions
-    })
+    for (const user of team) permissionOverwrites.push({ id: user.id, allow: permissions })
+    for (const user of users) permissionOverwrites.push({ id: user.id, allow: permissions })
 
     return permissionOverwrites
   }
@@ -156,11 +142,14 @@ export class TicketBuilder {
     return this
   }
 
-  async create (): Promise<Ticket | null | undefined> {
+  async create (): Promise<Ticket | null | undefined | void> {
     const { guild } = this.interaction
     const { category: categoryData } = this.options
-    if (guild === null || this.interaction instanceof Message) return
+    if (guild === null || this.interaction instanceof Message || this.interaction instanceof CommandInteraction) return
     if (!this.interaction.deferred) await this.interaction.deferReply({ ephemeral: true })
+
+    const templateData = await template.findOne({ where: { messageId: this.interaction.message?.id } })
+    if (templateData === null) return await new Error({ element: 'esse template', interaction: this.interaction }).notFound({ type: "Database" }).reply()
 
     const category = (await guild.channels.fetch()).find((channel) => channel?.type === ChannelType.GuildCategory && channel.name === categoryData.title)
         ?? await guild.channels.create({
@@ -185,11 +174,10 @@ export class TicketBuilder {
       messageId: messageMain.id,
     })
 
-    const ticketData = await ticket.create({ ...this.options, guild: { id: this.interaction.guildId }, template: { id: this.templateId } })
+    const ticketData = await ticket.create({ ...this.options, guild: { guildId: this.interaction.guildId }, template: { id: this.templateId } })
     const result = await ticket.save(ticketData) as Ticket
 
     if (result === null || result === undefined) {
-      console.error(result)
       await channel.delete('Error')
       await new Error({ element: 'salvar os dados no Database', interaction: this.interaction }).notPossible().reply()
       return null
@@ -216,35 +204,49 @@ export class TicketBuilder {
     return ticketData
   }
 
-  async load({ id }: { id: number }) {
-    const ticketData = await ticket.findOne({ where: { id } })
-    if (ticketData) this.options = ticketData
+  async loader() {
+    if (this.channelId === undefined) throw new Error({ element: 'executar essa ação, pois setTicket não foi configurado!', interaction: this.interaction }).notPossible().reply()
+
+    const ticketData = await ticket.findOne({ where: { channelId: this.channelId } })
+    
+    if (ticketData !== null) this.options = ticketData
+    return this
   }
 
-  async delete ({ id }: { id: number }) {
-    const ticketData = await ticket.findOne({ where: { id } })
+  async delete () {
+    if (this.interaction instanceof Message) return
+    const claimBuilder = new ClaimBuilder({ interaction: this.interaction })
+    const ticketData = await ticket.findOne({ where: { channelId: this.channelId }, relations: { claim: true } })
     if (ticketData === null) { await new Error({ element: 'as informações do ticket', interaction: this.interaction }).notFound({ type: 'Database' }).reply(); return }
     
     const channel = await this.interaction.client.channels.fetch(ticketData.channelId).catch(() => null)
     if (!channel?.isTextBased()) { await new Error({ element: ticketData.channelId, interaction: this.interaction }).notFound({ type: 'Channel' }).reply(); return }
-    
+
     await channel.delete()
     for (const { channelId, messageId } of ticketData.messages) {
       const channel = await this.interaction.client.channels.fetch(channelId).catch(() => null)
       if (channel === null || !channel.isTextBased()) continue
-      
+    
       const message = await channel.messages.fetch(messageId).catch(() => null)
       if (message === null) continue
 
       if (message.deletable) await message.delete()
     }
-    await ticket.delete({ id })
+    await ticket.delete({ id: ticketData.id })
+    await claimBuilder.delete(ticketData.claim.id)
   }
 
-  async edit ({ id }: { id: number }) { await ticket.update({ id }, this.options); return this }
+  async edit () { 
+    if (this.ticketId === undefined) throw new Error({ element: 'executar essa ação, pois o Id do ticket não foi setado!', interaction: this.interaction }).notPossible().reply()
 
-  async update ({ id }: { id: number }) {
-    const ticketData = await ticket.findOne({ where: { id } })
+    await ticket.update({ id: this.ticketId }, this.options)
+    return this
+  }
+
+  async update () {
+    if (this.ticketId === undefined) throw new Error({ element: 'executar essa ação, pois o Id do ticket não foi setado!', interaction: this.interaction }).notPossible().reply()
+
+    const ticketData = await ticket.findOne({ where: { id: this.ticketId } })
     if (ticketData === null) { await new Error({ element: 'as informações do ticket', interaction: this.interaction }).notFound({ type: 'Database' }).reply(); return }
     
     const channel = await this.interaction.client.channels.fetch(ticketData.channelId).catch(() => null)
@@ -257,7 +259,7 @@ export class TicketBuilder {
     const embed = this.embed as EmbedBuilder
     const buttons = this.buttons as ActionRowBuilder<ButtonBuilder>[]
 
-    await this.edit({ id })
+    await this.edit()
     await (channel as TextChannel).edit({ permissionOverwrites: this.permissions() })
     await message.edit({ embeds: [embed], components: buttons })
     return this
